refactor: separates out algos removing non useful ones
