+++ src/algorithms/array/__tests__/flatten.test.ts
@@ -1,0 +1,13 @@
+import { flatten } from '../flatten';
+
+describe('flatten', () => {
+    it('returns nested arrays flattened in one dimensional array', () => {
+        expect(flatten([1, [2], [3, [[4]]]])).toEqual([1, 2, 3, 4]);
+        expect(
+            flatten([
+                [1, [2]],
+                [3, [[4]]],
+            ]),
+        ).toEqual([1, 2, 3, 4]);
+    });
+});
