+++ src/algorithms/number/factorialize.ts
@@ -1,0 +1,4 @@
+// returns the factorial of passed integer
+export function factorialize(num: number): number {
+    return num < 1 ? 1 : num * factorialize(num - 1);
+}
