+++ src/algorithms/string/__tests__/convertHtmlEntities.test.ts
