+++ src/algorithms/string/__tests__/permutationCount.test.ts
@@ -1,0 +1,10 @@
+import { permutationCount } from '../permutationCount';
+
+describe('permutationCount', () => {
+    it('returns count of permutations without repeat characters', () => {
+        expect(permutationCount('a')).toEqual(1);
+        expect(permutationCount('aab')).toEqual(2);
+        expect(permutationCount('aabb')).toEqual(8);
+        expect(permutationCount('abcde')).toEqual(120);
+    });
+});
