+++ src/algorithms/string/convertBinaryToWords.ts
@@ -1,0 +1,7 @@
+// convert binary code to words
+export function convertBinaryToWords(str: string): string {
+    return str
+        .split(' ')
+        .map((binary) => String.fromCharCode(parseInt(binary, 2)))
+        .join('');
+}
