+++ src/algorithms/string/convertHtmlEntities.ts
